cpp: restructure cmake & headers to make it easier to use
- removes the need for libevents_definitions.h
- allows to install the library
- changes the protocol to be header-only to avoid the direct mavlink
  dependency when building the library
