+++ libs/cpp/Config.cmake.in
@@ -1,0 +1,4 @@
+
+@PACKAGE_INIT@
+
+include ( "${CMAKE_CURRENT_LIST_DIR}/libevents.cmake" )
