+++ libs/cpp/parse/CMakeLists.txt
@@ -1,3 +1,4 @@
+include(GNUInstallDirs)
 
 add_library(libevents_parser
 	parser.cpp
@@ -10,6 +11,22 @@ target_compile_options(libevents_parser
 	-Wconversion
 	-Wpedantic
 )
+set_target_properties(libevents_parser PROPERTIES PUBLIC_HEADER parser.h)
+install(TARGETS libevents_parser
+		EXPORT libevents
+		LIBRARY DESTINATION ${CMAKE_INSTALL_LIBDIR}
+		PUBLIC_HEADER DESTINATION ${CMAKE_INSTALL_INCLUDEDIR}/libevents/parse)
+
+add_library(nlohmann_json
+		INTERFACE
+		nlohmann_json/single_include/nlohmann/json_fwd.hpp
+)
+set_target_properties(nlohmann_json PROPERTIES PUBLIC_HEADER
+		nlohmann_json/single_include/nlohmann/json_fwd.hpp)
+install(TARGETS nlohmann_json
+		EXPORT libevents
+		LIBRARY DESTINATION ${CMAKE_INSTALL_LIBDIR}
+		PUBLIC_HEADER DESTINATION ${CMAKE_INSTALL_INCLUDEDIR}/libevents/parse/nlohmann_json/single_include/nlohmann)
 
 add_library(libevents_health_and_arming_checks
 	health_and_arming_checks.cpp
@@ -22,3 +39,8 @@ target_compile_options(libevents_health_and_arming_checks
 	-Wconversion
 	-Wpedantic
 )
+set_target_properties(libevents_health_and_arming_checks PROPERTIES PUBLIC_HEADER health_and_arming_checks.h)
+install(TARGETS libevents_health_and_arming_checks
+		EXPORT libevents
+		LIBRARY DESTINATION ${CMAKE_INSTALL_LIBDIR}
+		PUBLIC_HEADER DESTINATION ${CMAKE_INSTALL_INCLUDEDIR}/libevents/parse)
