+++ libs/cpp/parse/health_and_arming_checks.h
@@ -1,10 +1,9 @@
 #pragma once
 
-#include <libevents_definitions.h>
-
 #include <cstdint>
 #include <functional>
 
+#include "../common/event_type.h"
 #include "parser.h"
 
 namespace events
